delta_time is now in seconds instead of milliseconds

+++ include/mainmenu.h
@@ -27,7 +27,7 @@ namespace feed
         MainMenu(SDL_Surface* background, const glm::vec2& position);
 
         virtual void draw(SDL_Surface* screen) override final;
-        virtual void update(Uint32 delta_time) override final;
+        virtual void update(float delta_time) override final;
         virtual void handleSDLEvent(const SDL_Event& event) override final;
         virtual void handleMessage(const MessageQueue::Message& msg) override final;
