+++ src/mainmenu.cc
@@ -33,7 +33,7 @@ namespace feed
         main_.draw(screen);
     }
 
-    void MainMenu::update(Uint32 delta_time)
+    void MainMenu::update(float delta_time)
     {
 
     }
